requirements.txt
astroid==2.5.6
autopep8==1.5.7
black==21.9b0
click==8.0.1
Django==2.2
django-environ==0.8.1
django-seed==0.3.1
djangorestframework==3.12.4
Faker==9.5.1
flake8==3.9.2
Flask==2.0.1
greenlet==1.1.0
importlib-metadata==4.6.1
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.8.2
python-dateutil==2.8.2
pytz==2021.1
regex==2021.9.24
six==1.16.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
toposort==1.7
typed-ast==1.4.3
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
cffi==1.15.0 
cryptography==35.0.0 
django-rest-knox==4.1.0 
pycparser==2.20